+++ src/detection/gtk_qt/qt.c
@@ -1,7 +1,7 @@
 #include "fastfetch.h"
 #include "common/properties.h"
-#include "common/thread.h"
 #include "common/strutil.h"
+#include "common/FFcache.h"
 #include "detection/gtk_qt/gtk_qt.h"
 #include "detection/displayserver/displayserver.h"
 
@@ -172,38 +172,53 @@ static void detectKvantum(FFQtResult* result) {
                                                                });
 }
 
-const FFQtResult* ffDetectQt(void) {
-    static FFQtResult result;
+static FFQtResult result;
 
-    static bool init = false;
-    if (init) {
-        return &result;
-    }
-    init = true;
+static void initQtResult(void* storage) {
+    FFQtResult* qt = storage;
 
-    ffStrbufInit(&result.widgetStyle);
-    ffStrbufInit(&result.colorScheme);
-    ffStrbufInit(&result.icons);
-    ffStrbufInit(&result.font);
-    ffStrbufInit(&result.wallpaper);
+    ffStrbufInit(&qt->widgetStyle);
+    ffStrbufInit(&qt->colorScheme);
+    ffStrbufInit(&qt->icons);
+    ffStrbufInit(&qt->font);
+    ffStrbufInit(&qt->wallpaper);
 
     const FFDisplayServerResult* wmde = ffConnectDisplayServer();
 
     if (ffStrbufIgnCaseEqualS(&wmde->dePrettyName, FF_DE_PRETTY_PLASMA)) {
-        detectPlasma(&result);
+        detectPlasma(qt);
     } else if (ffStrbufIgnCaseEqualS(&wmde->dePrettyName, FF_DE_PRETTY_LXQT)) {
-        detectLXQt(&result);
+        detectLXQt(qt);
     } else {
         const char* qPlatformTheme = getenv("QT_QPA_PLATFORMTHEME");
         if (qPlatformTheme && (ffStrEquals(qPlatformTheme, "qt5ct") || ffStrEquals(qPlatformTheme, "qt6ct"))) {
-            detectQtCt(qPlatformTheme[2], &result);
+            detectQtCt(qPlatformTheme[2], qt);
         }
     }
 
-    if (ffStrbufEqualS(&result.widgetStyle, "kvantum") || ffStrbufEqualS(&result.widgetStyle, "kvantum-dark")) {
-        ffStrbufClear(&result.widgetStyle);
-        detectKvantum(&result);
+    if (ffStrbufEqualS(&qt->widgetStyle, "kvantum") || ffStrbufEqualS(&qt->widgetStyle, "kvantum-dark")) {
+        ffStrbufClear(&qt->widgetStyle);
+        detectKvantum(qt);
     }
+}
 
-    return &result;
+static void destroyQtResult(void* storage) {
+    FFQtResult* qt = storage;
+
+    ffStrbufDestroy(&qt->widgetStyle);
+    ffStrbufDestroy(&qt->colorScheme);
+    ffStrbufDestroy(&qt->icons);
+    ffStrbufDestroy(&qt->font);
+    ffStrbufDestroy(&qt->wallpaper);
+}
+
+static FFcacheEntry ffCacheEntryQt = {
+    .name = "qt",
+    .storage = &result,
+    .init = initQtResult,
+    .destroy = destroyQtResult,
+};
+
+const FFQtResult* ffDetectQt(void) {
+    return ffCacheGet(&ffCacheEntryQt);
 }